#559 - mu: break up vector monolith
