+++ src/mu/core/env.rs
@@ -1,13 +1,20 @@
 //  SPDX-FileCopyrightText: Copyright 2022 James M. Putnam (putnamjm.design@gmail.com)
 //  SPDX-License-Identifier: MIT
 
-//! env environment
-//!    Env
+//! env structure
+#[allow(unused_imports)]
 use {
     crate::{
-        core::{config::Config, frame::Frame, lib::Lib, types::Tag},
+        core::{
+            config::Config,
+            frame::Frame,
+            lib::Lib,
+            nursery::{Core as _, Nursery},
+            types::{Tag, TypeImage},
+        },
         images::bump_allocator::BumpAllocator,
-        types::{namespace::Namespace, vector::VecCacheMap},
+        types::namespace::Namespace,
+        vectors::cache::VecCacheMap,
         LIB,
     },
     cpu_time::ProcessTime,
@@ -24,6 +31,7 @@ pub struct Env {
 
     // heap
     pub heap: RwLock<BumpAllocator>,
+    pub nursery: RwLock<Nursery>,
     pub gc_root: RwLock<Vec<Tag>>,
     pub vector_map: RwLock<VecCacheMap>,
 
@@ -52,6 +60,7 @@ pub struct Env {
 impl Env {
     pub fn new(config: Config, _image: Option<Vec<u8>>) -> Self {
         let heap = BumpAllocator::new(config.npages, Tag::NTYPES);
+        let nursery = Nursery::new();
 
         let mut env = Env {
             config,
@@ -63,6 +72,7 @@ impl Env {
             lexical: RwLock::new(HashMap::new()),
             ns_map: RwLock::new(Vec::new()),
             null_ns: Tag::nil(),
+            nursery: RwLock::new(nursery),
             #[cfg(feature = "prof")]
             prof: RwLock::new(Vec::new()),
             #[cfg(feature = "prof")]
