+++ src/mu/core/gc.rs
@@ -71,6 +71,7 @@ impl Gc {
 
                 marked
             }
+            Tag::Nursery(_) => panic!(),
         }
     }
