+++ src/mu/features/prof.rs
@@ -20,8 +20,8 @@ use crate::{
         struct_::{Core as _, Struct},
         symbol::{Core as _, Symbol},
         vector::{Core as _, Vector},
-        vector_image::Core as _,
     },
+    vectors::image::Core as _,
 };
 use {futures::executor::block_on, futures_locks::RwLock};
