+++ src/mu/types/stream.rs
@@ -26,8 +26,8 @@ use crate::{
         fixnum::{Core as _, Fixnum},
         symbol::{Core as _, Symbol},
         vector::{Core as _, Vector},
-        vector_image::Core as _,
     },
+    vectors::core::Core as _,
 };
 
 use futures::executor::block_on;
